+++ pyproject.toml
@@ -1,0 +1,15 @@
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "program_youtube_downloader"
+version = "0.1.0"
+description = "CLI tool for downloading YouTube content"
+readme = "README.md"
+requires-python = ">=3.10"
+license = {text = "MIT"}
+dependencies = [
+    "pytubefix",
+    "colorama",
+]
